Add the ability to configure hazelcast under okapi
This adds options to set up hazelcast in order to provide memory sharing
across okapi pods for high availability.  Some of the functionality
requires creating a clusterrole, which end users might not have access
to do.  In that case there's an option to only set up the local role and
print out the clusterrole to send to your k8s administrators.